set all memory size parameters when creating buffers to size_t, to avoid problems for very big data sets. added some exception handling, but currently all openCL errors connected to a graphic drivers reset seem to crash mitk either way. setting the lower or upper slice bound updates now the other. made apodisation arrays persistant to avoid unneccesary recalculations.

+++ Modules/OpenCL/mitkOclDataSet.h
@@ -87,7 +87,7 @@ public:
   }
 
   /*! \brief returns the amount of elements in the DataSet */
-  int GetBufferSize() const
+  size_t GetBufferSize() const
   {
     return this->m_bufferSize;
   }
@@ -98,7 +98,7 @@ public:
   }
 
   /** @brief Set the amount of elements in buffer*/
-  void SetBufferSize(unsigned int size);
+  void SetBufferSize(size_t size);
 
   /** @brief Set the DataSet memory Size per Element in Bytes*/
   void SetBpE(unsigned short BpE);
@@ -118,7 +118,7 @@ protected:
 
 private:
   /*! GMEM Buffer Size in elements*/
-  unsigned int m_bufferSize;
+  size_t m_bufferSize;
 
   bool m_gpuModified;
   bool m_cpuModified;
